Rename BatchPredict Files
