+++ genai/batch_prediction/batchpredict_with_bq.py
