+++ genai/batch_prediction/batchpredict_with_gcs.py
